Fix hardware-test.bat
See https://github.com/mciantyre/teensy4-rs-template/commit/977065d5d506911c49642e85eba044126febb96a

+++ hardware-test.bat
@@ -10,7 +10,7 @@ IF NOT DEFINED scriptbuildflag (
 	SET scriptbuildmode=release
 )
 cargo build %scriptbuildflag% -p teensy4-examples --bin %scriptbin% || goto :error
-rmdir /S /Q out || goto :error
+if exist out rmdir /S /Q out || goto :error
 mkdir out || goto :error
 copy "target\thumbv7em-none-eabihf\%scriptbuildmode%\%scriptbin%" "out\%scriptbin%" || goto :error
 rust-objdump -d -S -C "out\%scriptbin%" > "out\%scriptbin%.lst" || goto :error
